% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{ACS2020_V2}
\alias{ACS2020_V2}
\title{ACS2020_V2}
\usage{
ACS2020_V2(
  SERV_DATA,
  RESPONDENTID,
  TOTALKCAL_ACS2020,
  VEG_SERV_ACS2020,
  VEG_ITEMS_SERV_ACS2020,
  FRT_SERV_ACS2020,
  FRT_ITEMS_SERV_ACS2020,
  WGRAIN_SERV_ACS2020,
  SSB_FRTJ_SERV_ACS2020,
  REDPROC_MEAT_SERV_ACS2020,
  HPFRG_SERV_ACS2020
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{TOTALKCAL_ACS2020}{The total kcal}

\item{VEG_SERV_ACS2020}{The serving size of All vegetable except potatoes and legume, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)}

\item{VEG_ITEMS_SERV_ACS2020}{The total number of vegetables (e.g., vegetable line items)  asked about on the survey/app}

\item{FRT_SERV_ACS2020}{The serving size of All whole fruits and no fruit juice, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g)}

\item{FRT_ITEMS_SERV_ACS2020}{The total number of fruits (e.g., fruit line items) asked about on that survey/app}

\item{WGRAIN_SERV_ACS2020}{The serving size of whole grains, unit=grams/day}

\item{SSB_FRTJ_SERV_ACS2020}{The serving size of sugar-sweetened beverages and non-100\% fruit juice, unit=servings/day = 1 ser= 8oz (1 oz. = 28.35 g)}

\item{REDPROC_MEAT_SERV_ACS2020}{The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)}

\item{HPFRG_SERV_ACS2020}{The daily servings of highly processed foods and refined grains per 1000 kcal(e.g., % calories from HPF and ref grains)}
}
\value{
The ACS2020_V2 index/score and its component scores
}
\description{
Calculate the American Cancer Society 2020 dietary index version 2 (using servings/d per 1000 kcal), using given the serving sizes of foods and nutrients consumed per 1 day. To help identify what are highly processed foods, here are some references: https://academic.oup.com/ajcn/article/101/6/1251/4626878, https://pubmed.ncbi.nlm.nih.gov/30744710/, http://archive.wphna.org/wp-content/uploads/2016/01/WN-2016-7-1-3-28-38-Monteiro-Cannon-Levy-et-al-NOVA.pdf
}
\examples{
ACS2020_V1(SERV_DATA, SERV_DATA$RESPONDENTID, SERV_DATA$TOTALKCAL_ACS2020, SERV_DATA$VEG_SERV_ACS2020, SERV_DATA$VEG_ITEMS_SERV_ACS2020, SERV_DATA$FRT_SERV_ACS2020, SERV_DATA$FRT_ITEMS_SERV_ACS2020, SERV_DATA$WGRAIN_SERV_ACS2020, SERV_DATA$SSB_FRTJ_SERV_ACS2020, SERV_DATA$REDPROC_MEAT_SERV_ACS2020, SERV_DATA$HPFRG_SERV_ACS2020)
}
