% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{AHEI}
\alias{AHEI}
\title{AHEI Calculation}
\usage{
AHEI(
  SERV_DATA,
  RESPONDENTID,
  VEG_SERV,
  FRT_SERV,
  WGRAIN_SERV,
  NUTSLEG_SERV,
  N3FAT_SERV,
  PUFA_SERV,
  SSB_FRTJ_SERV,
  REDPROC_MEAT_SERV,
  TRANS_SERV,
  SODIUM_SERV,
  ALCOHOL_SERV
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{VEG_SERV}{The serving size of All vegetable except potatoes and legume, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)}

\item{FRT_SERV}{The serving size of All whole fruits and no fruit juice, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g)}

\item{WGRAIN_SERV}{The serving size of whole grains, unit=grams/day}

\item{NUTSLEG_SERV}{The serving size of Nuts, legumes, and vegetable protein (e.g., tofu), unit=servings/day = 1 srv=1oz (28.35 g) of nuts and legume or 1 TBLSP peanut butter (15 mL), 1 cup legume = 4 oz}

\item{N3FAT_SERV}{The serving size of omega 3 fatty acid, unit=mg/day ( oz. = 28.35 g)}

\item{PUFA_SERV}{The serving size of PUFA, unit=\% of energy}

\item{SSB_FRTJ_SERV}{The serving size of sugar-sweetened beverages and non-100\% fruit juice, unit=servings/day = 1 ser= 8oz (1 oz. = 28.35 g)}

\item{REDPROC_MEAT_SERV}{The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)}

\item{TRANS_SERV}{The serving size of trans fat, unit=\% of energy}

\item{SODIUM_SERV}{The serving size of sodium, unit=mg/day}

\item{ALCOHOL_SERV}{The serving size of alcohol, including Wine, beer, "light" beer, liquor, unit=drink/day (12 oz. beer; 5 oz. wine; 1.5 oz. spirits) 1 oz = 28.35 g}
}
\value{
The AHEI index/score, AHEI
}
\description{
Calculate the AHEI dietary index, Alternative Healthy Eating Index, using given the serving sizes of foods and nutrients consumed per 1 day
}
\examples{
AHEI(SERV_DATA, RESPONDENTID, VEG_SERV, FRT_SERV, WGRAIN_SERV, NUTSLEG_SERV, N3FAT_SERV, PUFA_SERV, SSB_FRTJ_SERV, REDPROC_MEAT_SERV, TRANS_SERV,SODIUM_SERV, ALCOHOL_SERV)
}
