% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{AHEI_BLOCK}
\alias{AHEI_BLOCK}
\title{AHEI_BLOCK Calculation}
\usage{
AHEI_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw responses of the dietary assessment}
}
\value{
The AHEI and its component scores
}
\description{
Calculate the AHEI dietary index for the Block FFQ (2013) per 1 day
}
\examples{
AHEI_BLOCK(RAW_DATA)
}
