% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{AHEI_DHQ3}
\alias{AHEI_DHQ3}
\title{AHEI_DHQ3}
\usage{
AHEI_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Total Daily Results file, ending with results.csv}
}
\value{
The AHEI and its component scores
}
\description{
Calculate the AHEI for the DHQ3 data within 1 step
}
\examples{
DATA_PATH = "/Users/james/Desktop/data/results.csv"
AHEI_DHQ3(DATA_PATH)
}
