% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHEI_M_ASA24.R
\name{AHEI_M_ASA24}
\alias{AHEI_M_ASA24}
\title{AHEI_M_ASA24}
\usage{
AHEI_M_ASA24(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Items.csv}
}
\value{
The AHEI and its component scores
}
\description{
Calculate the AHEI (male only) for the ASA24 data within 1 step
}
\examples{
data("ASA24_exp_detailed")
AHEI_M_ASA24(ASA24_exp_detailed) # for participants who are all male
}
