% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{AHEIP_BLOCK}
\alias{AHEIP_BLOCK}
\title{AHEIP_BLOCK Calculation}
\usage{
AHEIP_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The AHEIP index/score and its components
}
\description{
Calculate the AHEIP dietary index (serving size based) per 1 day
}
\examples{
AHEIP_BLOCK(RAW_DATA)
}
