% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHEIP_DHQ3.R
\name{AHEIP_DHQ3}
\alias{AHEIP_DHQ3}
\title{AHEIP_DHQ3}
\usage{
AHEIP_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Detailed analysis file, ending with detail.csv}
}
\value{
The AHEIP and its component scores
}
\description{
Calculate the AHEIP (alternative healthy eating index - pregnancy) for the DHQ3 data within 1 step
}
\examples{
data("DHQ3_exp")
AHEIP_DHQ3(DHQ3_exp)
}
