% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{AHEIP_SERV}
\alias{AHEIP_SERV}
\title{AHEIP_SERV Calculation}
\usage{
AHEIP_SERV(RAW_DATA, TYPE)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}

\item{TYPE}{The type of dietary assessment you use. Current supported dietary assessment(s): BLOCK.}
}
\value{
The serving sizes for the AHEIP index/score
}
\description{
Calculate the serving sizes needed for calculating the AHEIP dietary index per 1 day
}
\examples{
AHEIP_SERV(RAW_DATA, TYPE="BLOCK")
}
