% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DASH_NHANES_FPED}
\alias{DASH_NHANES_FPED}
\title{DASH_NHANES_FPED}
\usage{
DASH_NHANES_FPED(FPED_PATH, NUTRIENT_PATH, DEMO_PATH, DBQ_PATH)
}
\arguments{
\item{FPED_PATH}{The file path for the FPED data. The file name should be like: fpre_dr1tot_1718.sas7bdat}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: DR1TOT_J.XPT}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}

\item{DBQ_PATH}{The file path for the Diet Behavior & Nutrition data. The file name should be like: DBQ_J.XPT}
}
\value{
The DASH and its component scores
}
\description{
Calculate the DASH for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
FPED_PATH = "/Users/james/Desktop/data/fpre_dr1tot_1718.sas7bdat"
NUTRIENT_PATH = "/Users/james/Desktop/data/DR1TOT_J.XPT"
DEMO_PATH = "/Users/james/Desktop/data/DEMO_J.XPT"
DBQ_PATH = "/Users/james/Desktop/data/DBQ_J.XPT"
DASH_NHANES_FPED(FPED_PATH, NUTRIENT_PATH, DEMO_PATH, DBQ_PATH)
}
