% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DASH_SERV}
\alias{DASH_SERV}
\title{DASH_SERV Calculation}
\usage{
DASH_SERV(RAW_DATA, TYPE)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}

\item{TYPE}{The type of dietary assessment you use. Current supported dietary assessment(s): BLOCK.}
}
\value{
The serving sizes for the DASH index/score
}
\description{
Calculate the serving sizes needed for calculating the DASH dietary index per 1 day
}
\examples{
DASH_SERV(RAW_DATA, TYPE="BLOCK")
}
