% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASHI.R
\name{DASHI}
\alias{DASHI}
\title{DASHI}
\usage{
DASHI(
  SERV_DATA,
  RESPONDENTID,
  TOTALKCAL_DASHI,
  VEG_SERV_DASHI,
  FRT_FRTJ_SERV_DASHI,
  NUTSLEG_SERV_DASHI,
  LOWF_DAIRY_SERV_DASHI,
  WGRAIN_SERV_DASHI,
  WHITEMEAT_SERV_DASHI,
  REDPROC_MEAT_SERV_DASHI,
  FATOIL_SERV_DASHI,
  SWEETS_SERV_DASHI,
  SODIUM_SERV_DASHI
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{TOTALKCAL_DASHI}{The total energy intake, unit = kcal/day per 2000 kcal}

\item{VEG_SERV_DASHI}{The serving size of All vegetable except potatoes and legume, unit=servings per 2000 kcal/day (1 serving = 1 cup of green leafy, 0.5 cup other vegetables (cooked or raw), 3/4 cup vegetable juice)}

\item{FRT_FRTJ_SERV_DASHI}{The serving size of All whole fruits + 100\% juice,  unit=servings per 2000 kcal/day (1 serving = 1 cup fruit, 1 cup fruit juice, 1/2 cup dried fruit)}

\item{NUTSLEG_SERV_DASHI}{The serving size of Nuts, legumes, and vegetable protein (e.g., tofu), unit=servings per 2000 kcal/day (1 serving =  1/3 cups of nuts, 2 tablespoons peanut butter, 1/2 cup cooked beans, 1/2 cup tofu)}

\item{LOWF_DAIRY_SERV_DASHI}{The serving size of low fat dairy, unit=servings per 2000 kcal/day (1 serving = 1 cup milk or yogurt, 1 1/2 oz natural cheese (e.g. Cheddar), 2 oz processed cheese (e.g. American))}

\item{WGRAIN_SERV_DASHI}{The serving size of whole grains, unit=servings per 2000 kcal/day (1 serving = 1oz)}

\item{WHITEMEAT_SERV_DASHI}{The serving size of poultry and fish, unit=servings per 2000 kcal/day (1 serving = 2.5oz)}

\item{REDPROC_MEAT_SERV_DASHI}{The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=servings per 2000 kcal/day (1 serving = 2.5oz)}

\item{FATOIL_SERV_DASHI}{The serving size of discretionary fats and oils, including added Plant oil + Animal fat, unit=servings per 2000 kcal/day (1 serving = 1 tablespoon or 14 g)}

\item{SWEETS_SERV_DASHI}{The serving size of sweets, unit=servings per 2000 kcal/day (1 serving = about 22 g)}

\item{SODIUM_SERV_DASHI}{The serving size of sodium, unit=mg per 2000 kcal/day}
}
\value{
The DASHI index/score
}
\description{
Calculate the DASHI dietary index (serving size-based), Dietary Approaches to Stop Hypertension, using given the serving sizes of foods and nutrients consumed per 1 day. All serving sizes should be divided by (total energy/2000 kcal) to adjust energy intake.
}
\examples{
DASHI(SERV_DATA, RESPONDENTID, SERV_DATA$TOTALKCAL_DASHI, SERV_DATA$VEG_SERV_DASHI, SERV_DATA$FRT_FRTJ_SERV_DASHI, SERV_DATA$NUTSLEG_SERV_DASHI, SERV_DATA$LOWF_DAIRY_SERV_DASHI, SERV_DATA$WGRAIN_SERV_DASHI, SERV_DATA$WHITEMEAT_SERV_DASHI, SERV_DATA$REDPROC_MEAT_SERV_DASHI, SERV_DATA$FATOIL_SERV_DASHI, SERV_DATA$SWEETS_SERV_DASHI, SERV_DATA$SODIUM_SERV_DASHI)
}
