% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DASHI_BLOCK}
\alias{DASHI_BLOCK}
\title{DASHI_BLOCK Calculation}
\usage{
DASHI_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The DASHI index/score and its components
}
\description{
Calculate the DASHI dietary index (serving size based) per 1 day
}
\examples{
DASHI_BLOCK(RAW_DATA)
}
