% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASHI_NHANES_FPED.R
\name{DASHI_NHANES_FPED}
\alias{DASHI_NHANES_FPED}
\title{DASHI_NHANES_FPED}
\usage{
DASHI_NHANES_FPED(NUTRIENT_PATH = NULL, NUTRIENT_PATH2 = NULL)
}
\arguments{
\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: DR1TOT_J.XPT}

\item{NUTRIENT_PATH2}{The file path for the NUTRIENT2 data. The file name should be like: DR2TOT_J.XPT}
}
\value{
The DASHI and its component scores and serving sizes
}
\description{
Calculate the DASHI (nutrient-based) for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
data("NHANES_20172018")
DASHI_NHANES_FPED(NUTRIENT_PATH = NHANES_20172018$NUTRIENT, NUTRIENT_PATH2 = NHANES_20172018$NUTRIENT2)
}
