% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DASHI_SERV}
\alias{DASHI_SERV}
\title{DASHI_SERV Calculation}
\usage{
DASHI_SERV(RAW_DATA, TYPE)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}

\item{TYPE}{The type of dietary assessment you use. Current supported dietary assessment(s): BLOCK.}
}
\value{
The serving sizes for the DASHI index/score
}
\description{
Calculate the serving sizes needed for calculating the DASHI dietary index per 1 day
}
\examples{
DASHI_SERV(RAW_DATA, TYPE="BLOCK")
}
