% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DI_GM.R
\name{DI_GM}
\alias{DI_GM}
\title{DI_GM Calculation}
\usage{
DI_GM(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  AVOCADO,
  BROCCOLI,
  CHICKPEA,
  COFFEE,
  CRANBERRY,
  FERMENTED_DAIRY,
  FIBER,
  GREEN_TEA,
  SOYBEAN,
  WHOLE_GRAIN,
  TOTAL_FAT_PERCENTAGE,
  PROCESSED_MEAT,
  RED_MEAT,
  REFINED_GRAIN
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{GENDER}{The gender of the participant}

\item{AVOCADO}{The serving size of Avocado. Unit does not matter here since the scoring is based on the median intake of the food in the population. However, it is important to keep the same unit for all indivudal food consumption data.}

\item{BROCCOLI}{The serving size of Broccoli. Unit does not matter here as indicated previously.}

\item{CHICKPEA}{The serving size of Chickpea. Unit does not matter here as indicated previously.}

\item{COFFEE}{The serving size of Coffee. Unit does not matter here as indicated previously.}

\item{CRANBERRY}{The serving size of Cranberry. Unit does not matter here as indicated previously.}

\item{FERMENTED_DAIRY}{The serving size of Fermented Dairy. Unit does not matter here as indicated previously.}

\item{FIBER}{The serving size of Fiber. Unit does not matter here as indicated previously.}

\item{GREEN_TEA}{The serving size of Green Tea. Unit does not matter here as indicated previously.}

\item{SOYBEAN}{The serving size of Soybean. Unit does not matter here as indicated previously.}

\item{WHOLE_GRAIN}{The serving size of Whole Grain. Unit does not matter here as indicated previously.}

\item{TOTAL_FAT_PERCENTAGE}{The total fat calorie percentage of the total calories consumed. Unit is percent calories.}

\item{PROCESSED_MEAT}{The serving size of Processed Meat. Unit does not matter here as indicated previously.}

\item{RED_MEAT}{The serving size of Red Meat. Unit does not matter here as indicated previously.}

\item{REFINED_GRAIN}{The serving size of Refined Grain. Unit does not matter here as indicated previously.}
}
\value{
The DI_GM index/score
}
\description{
Calculate the Dietary Index for Gut Microbiota (DI-GM), using given the serving sizes of foods and nutrients consumed per 1 day
}
\examples{
data("DI_GM_VALIDATION")
DI_GM(SERV_DATA = DI_GM_VALIDATION, RESPONDENTID = DI_GM_VALIDATION$id, GENDER = DI_GM_VALIDATION$gender, AVOCADO = DI_GM_VALIDATION$avocado, BROCCOLI = DI_GM_VALIDATION$broccoli, CHICKPEA = DI_GM_VALIDATION$chickpea, COFFEE = DI_GM_VALIDATION$coffee, CRANBERRY = DI_GM_VALIDATION$cranberry, FERMENTED_DAIRY = DI_GM_VALIDATION$fermented_dairy, FIBER = DI_GM_VALIDATION$fiber, GREEN_TEA = DI_GM_VALIDATION$green_tea, SOYBEAN = DI_GM_VALIDATION$soybean, WHOLE_GRAIN = DI_GM_VALIDATION$whole_grain, TOTAL_FAT_PERCENTAGE = DI_GM_VALIDATION$total_fat_percentage, PROCESSED_MEAT = DI_GM_VALIDATION$processed_meat, RED_MEAT = DI_GM_VALIDATION$red_meat, REFINED_GRAIN = DI_GM_VALIDATION$refined_grain)
}
