% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DI_GM_NHANES_FPED.R
\name{DI_GM_NHANES_FPED}
\alias{DI_GM_NHANES_FPED}
\title{DI_GM_NHANES_FPED}
\usage{
DI_GM_NHANES_FPED(
  FPED_IND_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  FPED_IND_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  AVOCADO_CODE = NULL,
  BROCCOLI_CODE = NULL,
  CHICKPEA_CODE = NULL,
  COFFEE_CODE = NULL,
  CRANBERRY_CODE = NULL,
  FERMENTED_DAIRY_CODE = NULL,
  GREEN_TEA_CODE = NULL,
  SOYBEAN_CODE = NULL
)
}
\arguments{
\item{FPED_IND_PATH}{The file path for the FPED IND data. The file name should be like: fped_dr1iff.sas7bdat}

\item{NUTRIENT_IND_PATH}{The file path for the NUTRIENT IND data. The file name should be like: DR1IFF_J}

\item{FPED_IND_PATH2}{The file path for the FPED IND data for day 2. The file name should be like: fped_dr1iff.sas7bdat}

\item{NUTRIENT_IND_PATH2}{The file path for the NUTRIENT IND data for day 2. The file name should be like: DR2IFF_J}

\item{AVOCADO_CODE}{The code for avocado in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{BROCCOLI_CODE}{The code for broccoli in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{CHICKPEA_CODE}{The code for chickpea in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{COFFEE_CODE}{The code for coffee in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{CRANBERRY_CODE}{The code for cranberry in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{GREEN_TEA_CODE}{The code for green tea in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{SOYBEAN_CODE}{The code for soybean foods in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{FERMENT_DAIRY_CODE}{The code for fermented dairy in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}
}
\value{
The DI_GM and its component scores and serving sizes
}
\description{
Calculate the Dietary Index for Gut Microbiota using a binary cutoff for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
data("NHANES_20172018")
DI_GM_NHANES_FPED(FPED_IND_PATH = NHANES_20172018$FPED_IND, NUTRIENT_IND_PATH = NHANES_20172018$NUTRIENT_IND, FPED_IND_PATH2 = NHANES_20172018$FPED_IND2, NUTRIENT_IND_PATH2 = NHANES_20172018$NUTRIENT_IND2)
}
