% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DI_GM_NHANES_MPED.R
\name{DI_GM_NHANES_MPED}
\alias{DI_GM_NHANES_MPED}
\title{DI_GM_NHANES_MPED}
\usage{
DI_GM_NHANES_MPED(
  MPED_PER_100_GRAM_PATH = NULL,
  WJFRT = NULL,
  DEMO_PATH,
  NUTRIENT_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  NUTRIENT_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  AVOCADO_CODE = NULL,
  BROCCOLI_CODE = NULL,
  CHICKPEA_CODE = NULL,
  COFFEE_CODE = NULL,
  CRANBERRY_CODE = NULL,
  FERMENTED_DAIRY_CODE = NULL,
  GREEN_TEA_CODE = NULL,
  SOYBEAN_CODE = NULL
)
}
\arguments{
\item{MPED_PER_100_GRAM_PATH}{The file path for the MPED per 100 gram data for the day 1 and day 2 data. The file name should be like: pyr_tot_d1.sas7bdat}

\item{WJFRT}{The file path for the WJFRT data for the day 1 and day2 data. The file name should be like: wjfrt.sas7bdat}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data for the day 1 data. The file name should be like: DR1TOT_J.XPT or DRXTOT_B.XPT}

\item{NUTRIENT_IND_PATH}{The file path for the NUTRIENT_IND data for the day 1 data. The file name should be like: DR1IFF_J.XPT}

\item{NUTRIENT_PATH2}{The file path for the NUTRIENT2 data for the day 2 data. The file name should be like: DR2TOT_J.XPT}

\item{NUTRIENT_IND_PATH2}{The file path for the NUTRIENT_IND2 data for the day 2 data The file name should be like: DR2IFF_J.XPT}

\item{AVOCADO_CODE}{The code for avocado in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{BROCCOLI_CODE}{The code for broccoli in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{CHICKPEA_CODE}{The code for chickpea in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{COFFEE_CODE}{The code for coffee in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{CRANBERRY_CODE}{The code for cranberry in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{GREEN_TEA_CODE}{The code for green tea in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{SOYBEAN_CODE}{The code for soybean foods in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}

\item{FERMENT_DAIRY_CODE}{The code for fermented dairy in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018}
}
\value{
The DI_GM and its component scores and serving sizes
}
\description{
Calculate the Dietary Index for Gut Microbiota using a binary cutoff for the NHANES_MPED data (before 2005, 1999-2004) within 1 step for day 1, day 2, or day 1 and 2 combined (age >= 2 only)
}
\examples{
data("NHANES_20032004")
DI_GM_NHANES_MPED(MPED_PER_100_GRAM_PATH = NHANES_20032004$MPED_PER_100_GRAM, WJFRT = NHANES_20032004$WJFRT, NUTRIENT_PATH = NHANES_20032004$NUTRIENT, NUTRIENT_IND_PATH = NHANES_20032004$NUTRIENT_IND, DEMO_PATH = NHANES_20032004$DEMO, NUTRIENT_PATH2 = NHANES_20032004$NUTRIENT2, NUTRIENT_IND_PATH2 = NHANES_20032004$NUTRIENT_IND2)
}
