% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII.R
\name{DII}
\alias{DII}
\title{DII}
\usage{
DII(
  SERV_DATA,
  RESPONDENTID,
  REPEATNUM = 1,
  ALCOHOL_DII = NULL,
  VITB12_DII = NULL,
  VITB6_DII = NULL,
  BCAROTENE_DII = NULL,
  CAFFEINE_DII = NULL,
  CARB_DII = NULL,
  CHOLES_DII = NULL,
  KCAL_DII = NULL,
  EUGENOL_DII = NULL,
  TOTALFAT_DII = NULL,
  FIBER_DII = NULL,
  FOLICACID_DII = NULL,
  GARLIC_DII = NULL,
  GINGER_DII = NULL,
  IRON_DII = NULL,
  MG_DII = NULL,
  MUFA_DII = NULL,
  NIACIN_DII = NULL,
  N3FAT_DII = NULL,
  N6FAT_DII = NULL,
  ONION_DII = NULL,
  PROTEIN_DII = NULL,
  PUFA_DII = NULL,
  RIBOFLAVIN_DII = NULL,
  SAFFRON_DII = NULL,
  SATFAT_DII = NULL,
  SE_DII = NULL,
  THIAMIN_DII = NULL,
  TRANSFAT_DII = NULL,
  TURMERIC_DII = NULL,
  VITA_DII = NULL,
  VITC_DII = NULL,
  VITD_DII = NULL,
  VITE_DII = NULL,
  ZN_DII = NULL,
  TEA_DII = NULL,
  FLA3OL_DII = NULL,
  FLAVONES_DII = NULL,
  FLAVONOLS_DII = NULL,
  FLAVONONES_DII = NULL,
  ANTHOC_DII = NULL,
  ISOFLAVONES_DII = NULL,
  PEPPER_DII = NULL,
  THYME_DII = NULL,
  ROSEMARY_DII = NULL
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{REPEATNUM}{The number of repeated record with each participant, 1st collection=1, 2nd collection =2, etc. If no repeat number is given, the default is 1}

\item{ALCOHOL_DII}{Unit=g}

\item{VITB12_DII}{Unit=μg}

\item{VITB6_DII}{Unit=mg}

\item{BCAROTENE_DII}{Unit=μg}

\item{CAFFEINE_DII}{Unit=g}

\item{CARB_DII}{Unit=g}

\item{CHOLES_DII}{Unit=mg}

\item{KCAL_DII}{Unit=KCAL_DII}

\item{EUGENOL_DII}{Unit=mg}

\item{TOTALFAT_DII}{Unit=g}

\item{FIBER_DII}{Unit=g}

\item{FOLICACID_DII}{Unit=μg}

\item{GARLIC_DII}{Unit=g}

\item{GINGER_DII}{Unit=g}

\item{IRON_DII}{Unit=mg}

\item{MG_DII}{Unit=mg}

\item{MUFA_DII}{Unit=g}

\item{NIACIN_DII}{Unit=mg}

\item{N3FAT_DII}{Unit=g}

\item{N6FAT_DII}{Unit=g}

\item{ONION_DII}{Unit=g}

\item{PROTEIN_DII}{Unit=g}

\item{PUFA_DII}{Unit=g}

\item{RIBOFLAVIN_DII}{Unit=mg}

\item{SAFFRON_DII}{Unit=g}

\item{SATFAT_DII}{Unit=g}

\item{SE_DII}{Unit=μg}

\item{THIAMIN_DII}{Unit=mg}

\item{TRANSFAT_DII}{Unit=g}

\item{TURMERIC_DII}{Unit=mg}

\item{VITA_DII}{Unit=RE}

\item{VITC_DII}{Unit=mg}

\item{VITD_DII}{Unit=μg}

\item{VITE_DII}{Unit=mg}

\item{ZN_DII}{Unit=mg}

\item{TEA_DII}{Unit=g}

\item{FLA3OL_DII}{Unit=mg}

\item{FLAVONES_DII}{Unit=mg}

\item{FLAVONOLS_DII}{Unit=mg}

\item{FLAVONONES_DII}{Unit=mg}

\item{ANTHOC_DII}{Unit=mg}

\item{ISOFLAVONES_DII}{Unit=mg}

\item{PEPPER_DII}{Unit=g}

\item{THYME_DII}{Unit=mg}

\item{ROSEMARY_DII}{Unit=mg}
}
\value{
The DII index/score
}
\description{
Calculate the DII dietary index, Dietary Inflammation Index, using given the serving sizes of foods and nutrients consumed per 1 day. Not all parameters are needed. You can give as many parameters as you have, and the results will only include scores for the parameters you included.
}
\examples{
data("DHQ3_exp")
DII(DHQ3_exp, DHQ3_exp$`Respondent ID`, 1, DHQ3_exp$`Alcohol (g)`, DHQ3_exp$`Vitamin B12 (mcg)`, DHQ3_exp$`Vitamin B6 (mg)`)
}
