% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_ASA24.R
\name{DII_ASA24}
\alias{DII_ASA24}
\title{DII_ASA24}
\usage{
DII_ASA24(DATA_PATH, RECALL_SUMMARIZE = TRUE)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Totals.csv.}

\item{RECALL_SUMMARIZE}{Whether to summarize the food group and nutrient intake over all days reported per individual per day. Default is TRUE.}
}
\value{
The DII and its component scores
}
\description{
Calculate the Dietary Inflammatory Index for the ASA24 data within 1 step
}
\examples{
data("ASA24_exp")
DII_ASA24(ASA24_exp, RECALL_SUMMARIZE = TRUE)
}
