% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DII_BLOCK}
\alias{DII_BLOCK}
\title{DII_BLOCK Calculation}
\usage{
DII_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The DII index/score and its components
}
\description{
Calculate the DII dietary index (serving size based) per 1 day
}
\examples{
DII_BLOCK(RAW_DATA)
}
