% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_DHQ3.R
\name{DII_DHQ3}
\alias{DII_DHQ3}
\title{DII_DHQ3}
\usage{
DII_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Total Daily Results file, ending with results.csv}
}
\value{
The DII and its component scores
}
\description{
Calculate the Dietary Inflammation Index (DII) for the DHQ3 data within 1 step
}
\examples{
data("DHQ3_exp")
DII_DHQ3(DHQ3_exp)
}
