% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_NHANES_MPED.R
\name{DII_NHANES_MPED}
\alias{DII_NHANES_MPED}
\title{DII_NHANES_MPED}
\usage{
DII_NHANES_MPED(
  MPED_PER_100_GRAM_PATH = NULL,
  WJFRT = NULL,
  NUTRIENT_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  DEMO_PATH,
  NUTRIENT_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL
)
}
\arguments{
\item{MPED_PER_100_GRAM_PATH}{The file path for the MPED per 100 gram data for the day 1 and day 2 data. The file name should be like: pyr_tot_d1.sas7bdat}

\item{WJFRT}{The file path for the WJFRT data for the day 1 and day2 data. The file name should be like: wjfrt.sas7bdat}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data for the day 1 data. The file name should be like: DR1TOT_J.XPT or DRXTOT_B.XPT}

\item{NUTRIENT_IND_PATH}{The file path for the NUTRIENT_IND data for the day 1 data. The file name should be like: DR1IFF_J.XPT}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}

\item{NUTRIENT_PATH2}{The file path for the NUTRIENT2 data for the day 2 data. The file name should be like: DR2TOT_J.XPT}

\item{NUTRIENT_IND_PATH2}{The file path for the NUTRIENT_IND2 data for the day 2 data The file name should be like: DR2IFF_J.XPT}
}
\value{
The DII and its component scores and serving sizes
}
\description{
Calculate the DII for the NHANES_MPED data (after 2005) within 1 step for day 1, day 2, or day 1 and 2 combined (age >= 2 only)
}
\examples{
data("NHANES_20032004")
DII_NHANES_MPED(MPED_PER_100_GRAM_PATH = NHANES_20032004$MPED_PER_100_GRAM, WJFRT = NHANES_20032004$WJFRT, NUTRIENT_PATH = NHANES_20032004$NUTRIENT, NUTRIENT_IND_PATH = NHANES_20032004$NUTRIENT_IND, DEMO_PATH = NHANES_20032004$DEMO, NUTRIENT_PATH2 = NHANES_20032004$NUTRIENT2, NUTRIENT_IND_PATH2 = NHANES_20032004$NUTRIENT_IND2)
}
