% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2015.R
\name{HEI2015}
\alias{HEI2015}
\title{HEI2015 Calculation}
\usage{
HEI2015(
  SERV_DATA,
  RESPONDENTID,
  TOTALKCAL_HEI2015,
  TOTALFRT_SERV_HEI2015,
  FRT_SERV_HEI2015,
  VEG_SERV_HEI2015,
  GREENNBEAN_SERV_HEI2015,
  TOTALPRO_SERV_HEI2015,
  SEAPLANTPRO_SERV_HEI2015,
  WHOLEGRAIN_SERV_HEI2015,
  DAIRY_SERV_HEI2015,
  FATTYACID_SERV_HEI2015,
  REFINEDGRAIN_SERV_HEI2015,
  SODIUM_SERV_HEI2015,
  ADDEDSUGAR_SERV_HEI2015,
  SATFAT_SERV_HEI2015
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{TOTALKCAL_HEI2015}{The total calorie from all foods and drinks}

\item{TOTALFRT_SERV_HEI2015}{The serving size of total fruits including fruit juice, unit= cup eq.}

\item{FRT_SERV_HEI2015}{The serving size of Citrus, Melons, Berries + Other Intact Fruits, unit= cup eq.}

\item{VEG_SERV_HEI2015}{The serving size of vegetables Total Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.}

\item{GREENNBEAN_SERV_HEI2015}{The serving size of Dark Green Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.}

\item{TOTALPRO_SERV_HEI2015}{The serving size of Total Meat, Poultry, and Seafood (including organ meats and cured meats) + Eggs + Nuts and Seeds + Soy + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz}

\item{SEAPLANTPRO_SERV_HEI2015}{The serving size of Seafood (high in n-3) + Seafood (low in n-3) + Soy + Nuts and Seeds + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz}

\item{WHOLEGRAIN_SERV_HEI2015}{The serving size of whole grains, unit=oz. eq.}

\item{DAIRY_SERV_HEI2015}{The serving size of all dairy, unit=cup eq.}

\item{FATTYACID_SERV_HEI2015}{The serving size of (Total Monounsaturated Fatty Acids + Total Polyunsaturated Fatty Acids)/Total Saturated Fatty Acids, unit=g}

\item{REFINEDGRAIN_SERV_HEI2015}{The serving size of refined grains, unit=oz. eq.}

\item{SODIUM_SERV_HEI2015}{The serving size of sodium, unit=mg}

\item{ADDEDSUGAR_SERV_HEI2015}{The serving size of added sugar, unit=tsp}

\item{SATFAT_SERV_HEI2015}{The serving size of Total Saturated Fatty Acids, unit=g}
}
\value{
The HEI2015 index/score
}
\description{
Calculate the HEI2015 dietary index, Healthy eating index 2015, using given the serving sizes of foods and nutrients consumed per 1 day
}
\examples{
data("HEI2015_VALIDATION")
HEI2015(SERV_DATA = HEI2015_VALIDATION, RESPONDENTID = HEI2015_VALIDATION$id, TOTALKCAL_HEI2015 = HEI2015_VALIDATION$kcal, TOTALFRT_SERV_HEI2015 = HEI2015_VALIDATION$total_fruit, FRT_SERV_HEI2015 = HEI2015_VALIDATION$whole_fruit, VEG_SERV_HEI2015 = HEI2015_VALIDATION$total_vegetable, GREENNBEAN_SERV_HEI2015 = HEI2015_VALIDATION$green_and_bean, TOTALPRO_SERV_HEI2015 = HEI2015_VALIDATION$total_protein, SEAPLANTPRO_SERV_HEI2015 = HEI2015_VALIDATION$seafood_plant_protein, WHOLEGRAIN_SERV_HEI2015 = HEI2015_VALIDATION$whole_grain, DAIRY_SERV_HEI2015 = HEI2015_VALIDATION$dairy, FATTYACID_SERV_HEI2015 = HEI2015_VALIDATION$fatty_acid, REFINEDGRAIN_SERV_HEI2015 = HEI2015_VALIDATION$refined_grain, SODIUM_SERV_HEI2015 = HEI2015_VALIDATION$sodium, ADDEDSUGAR_SERV_HEI2015 = HEI2015_VALIDATION$added_sugar, SATFAT_SERV_HEI2015 = HEI2015_VALIDATION$saturated_fat)
}
