% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{HEI2015_AARP}
\alias{HEI2015_AARP}
\title{HEI2015_AARP Calculation}
\usage{
HEI2015_AARP(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The HEI2015 index/score and its components
}
\description{
Calculate the HEI2015 dietary index (serving size based) per 1 day
}
\examples{
HEI2015_AARP(RAW_DATA)
}
