% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{HEI2015_NHANES_FPED}
\alias{HEI2015_NHANES_FPED}
\title{HEI2015_NHANES_FPED}
\usage{
HEI2015_NHANES_FPED(FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
\arguments{
\item{FPED_PATH}{The file path for the FPED data. The file name should be like: FPED.fped_dr1tot_1112.}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: NH.DR1TOT_G.}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: NH.DEMO_G.}
}
\value{
The HEI2015 and its component scores
}
\description{
Calculate the HEI2015 for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
FPED_PATH = "/Users/james/Desktop/data/FPED.fped_dr1tot_1112.csv"
NUTRIENT_PATH = "/Users/james/Desktop/data/NH.DR1TOT_G"
DEMO_PATH = "/Users/james/Desktop/data/NH.DEMO_G"
HEI2015_NHANES_FPED(FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
