% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2020.R
\name{HEI2020}
\alias{HEI2020}
\title{HEI2020 Calculation}
\usage{
HEI2020(
  SERV_DATA,
  RESPONDENTID,
  AGE,
  TOTALKCAL_HEI2020,
  TOTALFRT_SERV_HEI2020,
  FRT_SERV_HEI2020,
  VEG_SERV_HEI2020,
  GREENNBEAN_SERV_HEI2020,
  TOTALPRO_SERV_HEI2020,
  SEAPLANTPRO_SERV_HEI2020,
  WHOLEGRAIN_SERV_HEI2020,
  DAIRY_SERV_HEI2020,
  FATTYACID_SERV_HEI2020,
  REFINEDGRAIN_SERV_HEI2020,
  SODIUM_SERV_HEI2020,
  ADDEDSUGAR_SERV_HEI2020,
  SATFAT_SERV_HEI2020
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{AGE}{The age of the participant in years}

\item{TOTALKCAL_HEI2020}{The total calorie from all foods and drinks}

\item{TOTALFRT_SERV_HEI2020}{The serving size of total fruits including fruit juice, unit= cup eq.}

\item{FRT_SERV_HEI2020}{The serving size of Citrus, Melons, Berries + Other Intact Fruits, unit= cup eq.}

\item{VEG_SERV_HEI2020}{The serving size of vegetables Total Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.}

\item{GREENNBEAN_SERV_HEI2020}{The serving size of Dark Green Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.}

\item{TOTALPRO_SERV_HEI2020}{The serving size of Total Meat, Poultry, and Seafood (including organ meats and cured meats) + Eggs + Nuts and Seeds + Soy + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz}

\item{SEAPLANTPRO_SERV_HEI2020}{The serving size of Seafood (high in n-3) + Seafood (low in n-3) + Soy + Nuts and Seeds + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz}

\item{WHOLEGRAIN_SERV_HEI2020}{The serving size of whole grains, unit=oz. eq.}

\item{DAIRY_SERV_HEI2020}{The serving size of all dairy, unit=cup eq.}

\item{FATTYACID_SERV_HEI2020}{The serving size of (Total Monounsaturated Fatty Acids + Total Polyunsaturated Fatty Acids)/Total Saturated Fatty Acids, unit=g}

\item{REFINEDGRAIN_SERV_HEI2020}{The serving size of refined grains, unit=oz. eq.}

\item{SODIUM_SERV_HEI2020}{The serving size of sodium, unit=mg}

\item{ADDEDSUGAR_SERV_HEI2020}{The serving size of added sugar, unit=tsp}

\item{SATFAT_SERV_HEI2020}{The serving size of Total Saturated Fatty Acids, unit=g}
}
\value{
The HEI2020 index/score
}
\description{
Calculate the HEI2020 dietary index, Healthy eating index 2020 and HEI-Toddlers-2020, using given the serving sizes of foods and nutrients consumed per 1 day. This version has the Added sugar serving size as tsp (teaspoon) and the saturated fat serving size as g (gram), instead of the % of total energy for added sugar and saturated fat in the original HEI2020.
}
\examples{
data("HEI2020_VALIDATION")
HEI2020(SERV_DATA = HEI2020_VALIDATION, RESPONDENTID = HEI2020_VALIDATION$id, AGE = HEI2020_VALIDATION$age, TOTALKCAL_HEI2020 = HEI2020_VALIDATION$kcal, TOTALFRT_SERV_HEI2020 = HEI2020_VALIDATION$total_fruit, FRT_SERV_HEI2020 = HEI2020_VALIDATION$whole_fruit, VEG_SERV_HEI2020 = HEI2020_VALIDATION$total_vegetable, GREENNBEAN_SERV_HEI2020 = HEI2020_VALIDATION$green_and_bean, TOTALPRO_SERV_HEI2020 = HEI2020_VALIDATION$total_protein, SEAPLANTPRO_SERV_HEI2020 = HEI2020_VALIDATION$seafood_plant_protein, WHOLEGRAIN_SERV_HEI2020 = HEI2020_VALIDATION$whole_grain, DAIRY_SERV_HEI2020 = HEI2020_VALIDATION$dairy, FATTYACID_SERV_HEI2020 = HEI2020_VALIDATION$fatty_acid, REFINEDGRAIN_SERV_HEI2020 = HEI2020_VALIDATION$refined_grain, SODIUM_SERV_HEI2020 = HEI2020_VALIDATION$sodium, ADDEDSUGAR_SERV_HEI2020 = HEI2020_VALIDATION$added_sugar, SATFAT_SERV_HEI2020 = HEI2020_VALIDATION$saturated_fat)
}
