% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2020_ASA24.R
\name{HEI2020_ASA24}
\alias{HEI2020_ASA24}
\title{HEI2020_ASA24}
\usage{
HEI2020_ASA24(DATA_PATH, RECALL_SUMMARIZE = TRUE)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Totals.csv.}
}
\value{
The average HEI2020 and its component scores for each individual per day.
}
\description{
Calculate the HEI2020 for the ASA24 data within 1 step
}
\examples{
data("ASA24_exp")
HEI2020_ASA24(ASA24_exp, RECALL_SUMMARIZE = TRUE)
}
