% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2020_NHANES_FPED.R
\name{HEI2020_NHANES_FPED}
\alias{HEI2020_NHANES_FPED}
\title{HEI2020_NHANES_FPED}
\usage{
HEI2020_NHANES_FPED(
  FPED_PATH = NULL,
  NUTRIENT_PATH = NULL,
  DEMO_PATH,
  FPED_PATH2 = NULL,
  NUTRIENT_PATH2 = NULL
)
}
\arguments{
\item{FPED_PATH}{The file path for the FPED data. The file name should be like: fped_dr1tot_1112.sas7bdat}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: DR1TOT_J.XPT}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}

\item{FPED_PATH2}{The file path for the FPED2 data. The file name should be like: fped_DR2tot_1112.sas7bdat}

\item{NUTRIENT_PATH2}{The file path for the NUTRIENT2 data. The file name should be like: DR2TOT_J.XPT}
}
\value{
The HEI2020 and its component scores and serving sizes
}
\description{
Calculate the HEI2020 for the NHANES_FPED data (after 2005) within 1 step for day 1, day 2, or day 1 and 2 combined, including HEI2020 and HEI-Toddlers-2020
}
\examples{
data("NHANES_20172018")
HEI2020_NHANES_FPED(FPED_PATH = NHANES_20172018$FPED, NUTRIENT_PATH = NHANES_20172018$NUTRIENT, DEMO_PATH = NHANES_20172018$DEMO, FPED_PATH2 = NHANES_20172018$FPED2, NUTRIENT_PATH2 = NHANES_20172018$NUTRIENT2)
}
