% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2020_NHANES_MPED.R
\name{HEI2020_NHANES_MPED}
\alias{HEI2020_NHANES_MPED}
\title{HEI2020_NHANES_MPED}
\usage{
HEI2020_NHANES_MPED(
  MPED_PATH = NULL,
  NUTRIENT_PATH = NULL,
  DEMO_PATH,
  MPED_PATH2 = NULL,
  NUTRIENT_PATH2 = NULL
)
}
\arguments{
\item{MPED_PATH}{The file path for the MPED data for the day 1 data. The file name should be like: pyr_tot_d1.sas7bdat}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data for the day 1 data. The file name should be like: DR1TOT_J.XPT or DRXTOT_B.XPT}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}

\item{MPED_PATH2}{The file path for the MPED2 data for the day 2 data. The file name should be like: pyr_tot_d2.sas7bdat}

\item{NUTRIENT_PATH2}{The file path for the NUTRIENT2 data for the day 2 data. The file name should be like: DR2TOT_J.XPT}
}
\value{
The HEI2020 and its component scores and serving sizes
}
\description{
Calculate the HEI2020 for the NHANES_MPED data (after 2005) within 1 step for day 1, day 2, or day 1 and 2 combined, including HEI2020 and HEI-Toddlers-2020
}
\examples{
data("NHANES_20032004")
HEI2020_NHANES_MPED(MPED_PATH = NHANES_20032004$MPED, NUTRIENT_PATH = NHANES_20032004$NUTRIENT, DEMO_PATH = NHANES_20032004$DEMO, MPED_PATH2 = NHANES_20032004$MPED2, NUTRIENT_PATH2 = NHANES_20032004$NUTRIENT2)
}
