% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MED_ASA24.R
\name{MED_ASA24}
\alias{MED_ASA24}
\title{MED_ASA24}
\usage{
MED_ASA24(DATA_PATH, RECALL_SUMMARIZE = TRUE)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Totals.csv.}

\item{RECALL_SUMMARIZE}{Whether to summarize the food group and nutrient intake over all days reported per individual per day. Default is TRUE.}
}
\value{
The MED and its component scores
}
\description{
Calculate the MED for the ASA24 data within 1 step
}
\examples{
data("ASA24_exp")
MED_ASA24(ASA24_exp, RECALL_SUMMARIZE = TRUE)
}
