% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MED_BLOCK.R
\name{MED_BLOCK}
\alias{MED_BLOCK}
\title{MED_BLOCK Calculation}
\usage{
MED_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The MED index/score and its components
}
\description{
Calculate the MED dietary index (serving size based) per 1 day
}
\examples{
data("BLOCK_exp")
MED_BLOCK(BLOCK_exp)
}
