% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{MED_DHQ3}
\alias{MED_DHQ3}
\title{MED_DHQ3}
\usage{
MED_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Total Daily Results file, ending with results.csv}
}
\value{
The MED and its component scores
}
\description{
Calculate the MED for the DHQ3 data within 1 step
}
\examples{
DATA_PATH = "/Users/james/Desktop/data/results.csv"
MED_DHQ3(DATA_PATH)
}
