% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{MED_SERV}
\alias{MED_SERV}
\title{MED_SERV Calculation}
\usage{
MED_SERV(RAW_DATA, TYPE)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}

\item{TYPE}{The type of dietary assessment you use. Current supported dietary assessment(s): BLOCK.}
}
\value{
The serving sizes for the MED index/score
}
\description{
Calculate the serving sizes needed for calculating the MED dietary index per 1 day
}
\examples{
MED_SERV(RAW_DATA, TYPE="BLOCK")
}
