% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{MEDI_BLOCK}
\alias{MEDI_BLOCK}
\title{MEDI_BLOCK Calculation}
\usage{
MEDI_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The MEDI index/score and its components
}
\description{
Calculate the MEDI dietary index (serving size based) per 1 day
}
\examples{
MEDI_BLOCK(RAW_DATA)
}
