% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDI_NHANES_FPED.R
\name{MEDI_NHANES_FPED}
\alias{MEDI_NHANES_FPED}
\title{MEDI_NHANES_FPED}
\usage{
MEDI_NHANES_FPED(FPED_IND_PATH, NUTRIENT_IND_PATH)
}
\arguments{
\item{FPED_IND_PATH}{The file path for the FPED IND data. The file name should be like: fped_dr1iff.sas7bdat}

\item{NUTRIENT_IND_PATH}{The file path for the NUTRIENT IND data. The file name should be like: DR1IFF_J}
}
\value{
The MEDI and its component scores and serving sizes
}
\description{
Calculate the MEDI (serving size-based) using a binary cutoff for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
data("NHANES_20172018")
MEDI_NHANES_FPED(NHANES_20172018$FPED_IND, NHANES_20172018$NUTRIENT_IND)
}
