% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{NHANES_FPED_PRE_HEI15}
\alias{NHANES_FPED_PRE_HEI15}
\title{NHANES_FPED_PRE_HEI15}
\usage{
NHANES_FPED_PRE_HEI15(FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
\arguments{
\item{FPED_PATH}{The file path for the FPED data. The file name should be like: FPED.fped_dr1tot_1112.}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: NH.DR1TOT_G.}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: NH.DEMO_G.}
}
\value{
The cleaned data ready for calculating the serving sizes for HEI2015.
}
\description{
Prepare the NHANES_FPED data (after 2005) for calculating the serving sizes for HEI2015.
}
\examples{
FPED_PATH = "/Users/james/Desktop/data/FPED.fped_dr1tot_1112.csv"
NUTRIENT_PATH = "/Users/james/Desktop/data/NH.DR1TOT_G"
DEMO_PATH = "/Users/james/Desktop/data/NH.DEMO_G"
NHANES_FPED_PRE_HEI15 (FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
