% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\docType{data}
\name{STD_FOOD_FREQ}
\alias{STD_FOOD_FREQ}
\title{AHEI_SERV Calculation}
\format{
An object of class \code{numeric} of length 9.
}
\usage{
STD_FOOD_FREQ
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}

\item{TYPE}{The type of dietary assessment you use. Current supported dietary assessment(s): BLOCK.}
}
\value{
The serving sizes for the AHEI index/score
}
\description{
Calculate the serving sizes needed for calculating the AHEI dietary index per 1 day
}
\examples{
AHEI_SERV(RAW_DATA, TYPE="BLOCK")
}
\keyword{datasets}
