% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACS2020_V2.R
\name{ACS2020_V2}
\alias{ACS2020_V2}
\title{ACS2020_V2}
\usage{
ACS2020_V2(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  TOTALKCAL_ACS2020,
  VEG_SERV_ACS2020,
  VEG_ITEMS_SERV_ACS2020,
  FRT_SERV_ACS2020,
  FRT_ITEMS_SERV_ACS2020,
  WGRAIN_SERV_ACS2020,
  REDPROC_MEAT_SERV_ACS2020,
  HPFRG_SERV_ACS2020,
  SSB_FRTJ_SERV_ACS2020
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{GENDER}{The gender for each participant, 1=male, and 2=female}

\item{TOTALKCAL_ACS2020}{The total kcal}

\item{VEG_SERV_ACS2020}{The serving size of All vegetable except potatoes and starchy vegetable, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)}

\item{VEG_ITEMS_SERV_ACS2020}{The total number of unique vegetables (e.g., vegetable line items) reported by the participant. For example, a report of 3 servings of lettuce, 1 serving of kale, and 0.5 servings of broccoli would count as 3 vegetable line items.}

\item{FRT_SERV_ACS2020}{The serving size of All whole fruits and no fruit juice, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit; 0.5 medium avocado)}

\item{FRT_ITEMS_SERV_ACS2020}{The number of unique fruits (e.g., fruit line items) asked about on that survey/app (measuring the variety of fruits). For example, 3 serving of apple, 3 servings of banana, and 3 servings of blueberry are just 3 total number line items}

\item{WGRAIN_SERV_ACS2020}{The serving size of whole grains, unit=grams/day}

\item{REDPROC_MEAT_SERV_ACS2020}{The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)}

\item{HPFRG_SERV_ACS2020}{The daily servings of highly processed foods and refined grains per 1000 kcal, note: the ultra-processed variable for the score should not double count foods included in other parts of the score, for example, sugar-sweetened beverages or processed meats}

\item{SSB_FRTJ_SERV_ACS2020}{The serving size of sugar-sweetened beverages and non-100\% fruit juice, unit=servings/day = 1 ser= 8oz (1 oz. = 28.35 g)}
}
\value{
The ACS2020_V2 index/score and its component scores
}
\description{
Calculate the American Cancer Society 2020 dietary index version 2 (using servings/d per 1000 kcal), using given the serving sizes of foods and nutrients consumed per 1 day.
}
