% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHEI_DHQ3.R
\name{AHEI_DHQ3}
\alias{AHEI_DHQ3}
\title{AHEI_DHQ3}
\usage{
AHEI_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Detailed analysis file, ending with detail.csv}
}
\value{
The AHEI and its component scores
}
\description{
Calculate the AHEI for the DHQ3 data within 1 step
}
\examples{
data("DHQ3_exp_detailed")
AHEI_DHQ3(DHQ3_exp_detailed)
}
