% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHEI_NHANES_FPED.R
\name{AHEI_NHANES_FPED}
\alias{AHEI_NHANES_FPED}
\title{AHEI_NHANES_FPED}
\usage{
AHEI_NHANES_FPED(
  FPED_IND_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  FPED_IND_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  SSB_code = NULL
)
}
\arguments{
\item{FPED_IND_PATH}{The file path for the FPED IND data for day 1. The file name should be like: fped_dr1iff.sas7bdat}

\item{NUTRIENT_IND_PATH}{The file path for the NUTRIENT IND data for day 1. The file name should be like: DR1IFF_J}

\item{FPED_IND_PATH2}{The file path for the FPED IND data for day 2. The file name should be like: fped_dr2iff.sas7bdat}

\item{NUTRIENT_IND_PATH2}{The file path for the NUTRIENT IND data for day 2. The file name should be like: DR2IFF_J}

\item{SSB_code}{The food code for sugar sweetened beverage, default is the SSB code from 17-18 FNDDS file.}
}
\value{
The AHEI and its component scores and serving sizes
}
\description{
Calculate the AHEI for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
data("NHANES_20172018")
AHEI_NHANES_FPED(NHANES_20172018$FPED_IND, NHANES_20172018$NUTRIENT_IND, NHANES_20172018$FPED_IND2, NHANES_20172018$NUTRIENT_IND2)
}
