% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASH_ASA24.R
\name{DASH_ASA24}
\alias{DASH_ASA24}
\title{DASH_ASA24}
\usage{
DASH_ASA24(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Items.csv}
}
\value{
The DASH and its component scores
}
\description{
Calculate the DASH for the ASA24 data within 1 step
}
\examples{
data("ASA24_exp_detailed")
DASH_ASA24(ASA24_exp_detailed)
}
