% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DASH_BLOCK}
\alias{DASH_BLOCK}
\title{DASH_BLOCK Calculation}
\usage{
DASH_BLOCK(RAW_DATA)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}
}
\value{
The DASH and its component scores
}
\description{
Calculate the DASH dietary index for Block FFQ (2013) per 1 day
}
\examples{
DASH_BLOCK(RAW_DATA)
}
