% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASH_DHQ3.R
\name{DASH_DHQ3}
\alias{DASH_DHQ3}
\title{DASH_DHQ3}
\usage{
DASH_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Detailed analysis file, ending with detail.csv}
}
\value{
The DASH and its component scores
}
\description{
Calculate the DASH for the DHQ3 data within 1 step
}
\examples{
DATA_PATH = "/Users/james/Desktop/data/detail.csv"
DASH_DHQ3(DATA_PATH)
}
