% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASH_NHANES_FPED.R
\name{DASH_NHANES_FPED}
\alias{DASH_NHANES_FPED}
\title{DASH_NHANES_FPED}
\usage{
DASH_NHANES_FPED(
  FPED_IND_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  FPED_IND_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  SKIM_MILK_code = NULL,
  LOWF_CHEESE_code = NULL,
  SSB_code = NULL
)
}
\arguments{
\item{FPED_IND_PATH}{The file path for the FPED IND data. The file name should be like: fped_dr1iff.sas7bdat}

\item{NUTRIENT_IND_PATH}{The file path for the NUTRIENT IND data. The file name should be like: DR1IFF_J}

\item{FPED_IND_PATH2}{The file path for the FPED IND data for day 2. The file name should be like: fped_dr2iff.sas7bdat}

\item{NUTRIENT_IND_PATH2}{The file path for the NUTRIENT IND data for day 2. The file name should be like: DR2IFF_J}

\item{SKIM_MILK_code}{The food code for skim milk. The default food codes are from 17-18 FNDDS file.}

\item{LOWF_CHEESE_code}{The food code for low-fat cheese. The default food codes are from 17-18 FNDDS file.}

\item{SSB_code}{The food code for sugar-sweetened beverages (SSB). The default food codes are from 17-18 FNDDS file.}
}
\value{
The DASH and its component scores and serving sizes
}
\description{
Calculate the DASH for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
data("NHANES_20172018")
DASH_NHANES_FPED(NHANES_20172018$FPED_IND, NHANES_20172018$NUTRIENT_IND, NHANES_20172018$FPED_IND2, NHANES_20172018$NUTRIENT_IND2)
}
