% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DASHI}
\alias{DASHI}
\title{DASHI Calculation}
\usage{
DASHI(
  SERV_DATA,
  VEG_SERV,
  FRT_SERV,
  NUTSLEG_SERV,
  LOWFATDAIRY_SERV,
  WGRAIN_SERV,
  ALLMEAT_SERV,
  REDPROC_MEAT_SERV,
  FATOIL_SERV,
  ADDEDSUGAR_SERV,
  SODIUM_SERV
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{VEG_SERV}{The serving size of All vegetable except potatoes and legume, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)}

\item{FRT_SERV}{The serving size of All whole fruits + 100\% juice,  unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g); 1 cup fruit juice}

\item{NUTSLEG_SERV}{The serving size of Nuts, legumes, and vegetable protein (e.g., tofu), unit=servings/day = 1 srv=1oz (28.35 g) of nuts or 1 TBLSP peanut butter (15 mL), 1 cup legume = 4 oz}

\item{LOWFATDAIRY_SERV}{The serving size of low fat dairy, including 2\% or less fat milk + yogurt + low-fat ice cream and frozen yogurt + low-fat cheese, unit=servings/day = 1 glass milk + 1 cup yogurt + 1/2 cup ice cream/frozen yogurt + 1 slice cheese}

\item{WGRAIN_SERV}{The serving size of whole grains, unit=1oz}

\item{ALLMEAT_SERV}{The serving size of all meat consumption, including meat, fish, and poultry, unit=servings/day = 1oz}

\item{REDPROC_MEAT_SERV}{The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)}

\item{FATOIL_SERV}{The serving size of discretionary fats and oils, including added Plant oil + Animal fat, unit=servings/day = 1tbsp = 14 g}

\item{SODIUM_SERV}{The serving size of sodium, unit=mg/day}
}
\value{
The DASHI index/score
}
\description{
Calculate the DASHI dietary index (serving size-based), Dietary Approaches to Stop Hypertension, using given the serving sizes of foods and nutrients consumed per 1 day
}
\examples{
DASHI = function(SERV_DATA, VEG_SERV, FRT_SERV, NUTSLEG_SERV, LOWFATDAIRY_SERV, WGRAIN_SERV, ALLMEAT_SERV, REDPROC_MEAT_SERV, FATOIL_SERV, ADDEDSUGAR_SERV, SODIUM_SERV)
}
