% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASHI.R
\name{DASHI}
\alias{DASHI}
\title{DASHI}
\usage{
DASHI(
  SERV_DATA,
  RESPONDENTID,
  TOTALKCAL_DASHI,
  TOTAL_FAT_DASHI,
  SAT_FAT_DASHI,
  PROTEIN_DASHI,
  CHOLESTEROL_DASHI,
  FIBER_DASHI,
  POTASSIUM_DASHI,
  MAGNESIUM_DASHI,
  CALCIUM_DASHI,
  SODIUM_DASHI
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{TOTALKCAL_DASHI}{The total energy intake, unit = kcal/day}

\item{TOTAL_FAT_DASHI}{The total fat intake, unit = g/day}

\item{SAT_FAT_DASHI}{The saturated fat intake, unit = g/day}

\item{PROTEIN_DASHI}{The protein intake, unit = g/day}

\item{CHOLESTEROL_DASHI}{The cholesterol intake, unit = mg/day}

\item{FIBER_DASHI}{The fiber intake, unit = g/day}

\item{POTASSIUM_DASHI}{The potassium intake, unit = mg/day}

\item{MAGNESIUM_DASHI}{The magnesium intake, unit = mg/day}

\item{CALCIUM_DASHI}{The calcium intake, unit = mg/day}

\item{SODIUM_DASHI}{The sodium intake, unit = mg/day}
}
\value{
The DASHI index/score
}
\description{
Calculate the DASHI dietary index (nutrient-based), Dietary Approaches to Stop Hypertension, using given the nutrients consumed per 1 day. All nutrients will be divided by (total energy/2000 kcal) to adjust energy intake.
}
\examples{
data("DASH_trial")
DASHI(SERV_DATA = DASH_trial, RESPONDENTID = DASH_trial$Diet_Type, TOTALKCAL_DASHI = DASH_trial$Kcal, TOTAL_FAT_DASHI = DASH_trial$Totalfat_Percent, SAT_FAT_DASHI = DASH_trial$Satfat_Percent, PROTEIN_DASHI = DASH_trial$Protein_Percent, CHOLESTEROL_DASHI = DASH_trial$Cholesterol, FIBER_DASHI = DASH_trial$Fiber, POTASSIUM_DASHI = DASH_trial$Potassium, MAGNESIUM_DASHI = DASH_trial$Magnesium, CALCIUM_DASHI = DASH_trial$Calcium, SODIUM_DASHI = DASH_trial$Sodium)
}
