% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DII}
\alias{DII}
\title{DII}
\usage{
DII(
  SERV_DATA,
  RESPONDENTID,
  ALCOHOL = NULL,
  VITB12 = NULL,
  VITB6 = NULL,
  BCAROTENE = NULL,
  CAFFEINE = NULL,
  CARB = NULL,
  CHOLES = NULL,
  KCAL = NULL,
  EUGENOL = NULL,
  TOTALFAT = NULL,
  FIBER = NULL,
  FOLICACID = NULL,
  GARLIC = NULL,
  GINGER = NULL,
  IRON = NULL,
  MG = NULL,
  MUFA = NULL,
  NIACIN = NULL,
  N3FAT = NULL,
  N6FAT = NULL,
  ONION = NULL,
  PROTEIN = NULL,
  PUFA = NULL,
  RIBOFLAVIN = NULL,
  SAFFRON = NULL,
  SATFAT = NULL,
  SE = NULL,
  THIAMIN = NULL,
  TRANSFAT = NULL,
  TURMERIC = NULL,
  VITA = NULL,
  VITC = NULL,
  VITD = NULL,
  VITE = NULL,
  ZN = NULL,
  TEA = NULL,
  FLA3OL = NULL,
  FLAVONES = NULL,
  FLAVONOLS = NULL,
  FLAVONONES = NULL,
  ANTHOC = NULL,
  ISOFLAVONES = NULL,
  PEPPER = NULL,
  THYME = NULL,
  ROSEMARY = NULL
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{RESPONDENTID}{The unique participant ID for each participant}

\item{ALCOHOL}{Unit=g}

\item{VITB12}{Unit=μg}

\item{VITB6}{Unit=mg}

\item{CAFFEINE}{Unit=g}

\item{EUGENOL}{Unit=mg}

\item{FIBER}{Unit=g}

\item{FOLICACID}{Unit=μg}

\item{GARLIC}{Unit=g}

\item{GINGER}{Unit=g}

\item{IRON}{Unit=mg}

\item{MUFA}{Unit=g}

\item{NIACIN}{Unit=mg}

\item{ONION}{Unit=g}

\item{PROTEIN}{Unit=g}

\item{PUFA}{Unit=g}

\item{RIBOFLAVIN}{Unit=mg}

\item{SAFFRON}{Unit=g}

\item{SATFAT}{Unit=g}

\item{THIAMIN}{Unit=mg}

\item{TRANSFAT}{Unit=g}

\item{TURMERIC}{Unit=mg}

\item{VITA}{Unit=RE}

\item{VITC}{Unit=mg}

\item{VITD}{Unit=μg}

\item{VITE}{Unit=mg}

\item{TEA}{Unit=g}

\item{FLAVONES}{Unit=mg}

\item{FLAVONOLS}{Unit=mg}

\item{FLAVONONES}{Unit=mg}

\item{ISOFLAVONES}{Unit=mg}

\item{PEPPER}{Unit=g}

\item{THYME}{Unit=mg}

\item{ROSEMARY}{Unit=mg}

\item{BETACAROTENE}{Unit=μg}

\item{CARBOHYDRATE}{Unit=g}

\item{CHOLESTEROL}{Unit=mg}

\item{ENERGY}{Unit=kcal}

\item{FAT}{Unit=g}

\item{MAGNESIUM}{Unit=mg}

\item{N3FATTYACID}{Unit=g}

\item{N6FATTYACID}{Unit=g}

\item{SELENIUM}{Unit=μg}

\item{ZINC}{Unit=mg}

\item{FLAVAN3OL}{Unit=mg}

\item{ANTHOCYANIDINS}{Unit=mg}
}
\value{
The DII index/score
}
\description{
Calculate the DII dietary index, Dietary Inflammation Index, using given the serving sizes of foods and nutrients consumed per 1 day. Not all parameters are needed. You can give as many parameters as you have, and the results will only include scores for the parameters you included.
}
\examples{
DII(SERV_DATA, SERV_DATA$RESPONDENTID, SERV_DATA$ALCOHOL, SERV_DATA$VITB12, SERV_DATA$VITB6, SERV_DATA$BETACAROTENE, SERV_DATA$CAFFEINE, SERV_DATA$CARBOHYDRATE, SERV_DATA$CHOLESTEROL, SERV_DATA$ENERGY, SERV_DATA$EUGENOL, SERV_DATA$FAT, SERV_DATA$FIBER, SERV_DATA$FOLICACID, SERV_DATA$GARLIC, SERV_DATA$GINGER, SERV_DATA$IRON, SERV_DATA$MAGNESIUM, SERV_DATA$MUFA, SERV_DATA$NIACIN, SERV_DATA$N3FATTYACID, SERV_DATA$N6FATTYACID, SERV_DATA$ONION, SERV_DATA$PROTEIN, SERV_DATA$PUFA, SERV_DATA$RIBOFLAVIN, SERV_DATA$SAFFRON, SERV_DATA$SATFAT, SERV_DATA$SELENIUM, SERV_DATA$THIAMIN, SERV_DATA$TRANSFAT, SERV_DATA$TURMERIC, SERV_DATA$VITA, SERV_DATA$VITC, SERV_DATA$VITD, SERV_DATA$VITE, SERV_DATA$ZINC, SERV_DATA$TEA, SERV_DATA$FLAVAN3OL, SERV_DATA$FLAVONES, SERV_DATA$FLAVONOLS, SERV_DATA$FLAVONONES, SERV_DATA$ANTHOCYANIDINS, SERV_DATA$ISOFLAVONES, SERV_DATA$PEPPER, SERV_DATA$THYME, SERV_DATA$ROSEMARY)
}
