% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{DII_ASA24}
\alias{DII_ASA24}
\title{DII_ASA24}
\usage{
DII_ASA24(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Totals.csv.}
}
\value{
The DII and its component scores
}
\description{
Calculate the Dietary Inflammatory Index for the ASA24 data within 1 step
}
\examples{
DATA_PATH = "/Users/james/Desktop/data/Totals.csv"
DII_ASA24(DATA_PATH)
}
