% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2015_ASA24.R
\name{HEI2015_ASA24}
\alias{HEI2015_ASA24}
\title{HEI2015_ASA24}
\usage{
HEI2015_ASA24(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The file name should be like: Totals.csv.}
}
\value{
The average HEI2015 and its component scores for each individual per day.
}
\description{
Calculate the HEI2015 for the ASA24 data within 1 step
}
\examples{
data("ASA24_exp")
HEI2015_ASA24(ASA24_exp)
}
