% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2015_DHQ3.R
\name{HEI2015_DHQ3}
\alias{HEI2015_DHQ3}
\title{HEI2015_DHQ3}
\usage{
HEI2015_DHQ3(DATA_PATH)
}
\arguments{
\item{DATA_PATH}{The file path for the data. The data is Total Daily Results file, ending with results.csv}
}
\value{
The HEI2015 and its component scores
}
\description{
Calculate the HEI2015 for the DHQ3 data within 1 step
}
\examples{
data("DHQ3_exp")
HEI2015_DHQ3(DHQ3_exp)
}
