% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2015_NHANES_FPED.R
\name{HEI2015_NHANES_FPED}
\alias{HEI2015_NHANES_FPED}
\title{HEI2015_NHANES_FPED}
\usage{
HEI2015_NHANES_FPED(FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
\arguments{
\item{FPED_PATH}{The file path for the FPED data. The file name should be like: fped_dr1tot_1112.sas7bdat}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: DR1TOT_J.XPT}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}
}
\value{
The AHEI and its component scores and serving sizes
}
\description{
Calculate the HEI2015 for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
data("NHANES_20172018")
HEI2015_NHANES_FPED(NHANES_20172018$FPED, NHANES_20172018$NUTRIENT, NHANES_20172018$DEMO)
}
