% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{HEI2015_SERV}
\alias{HEI2015_SERV}
\title{HEI2015_SERV Calculation}
\usage{
HEI2015_SERV(RAW_DATA, TYPE)
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw data of the dietary assessment}

\item{TYPE}{The type of dietary assessment you use. Current supported dietary assessment(s): BLOCK.}
}
\value{
The serving sizes for the HEI2015 index/score
}
\description{
Calculate the serving sizes needed for calculating the HEI2015 dietary index per 1 day
}
\examples{
HEI2015_SERV(RAW_DATA, TYPE="BLOCK")
}
