% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{MED}
\alias{MED}
\title{MED Calculation}
\usage{
MED(
  SERV_DATA,
  FRT_FRTJ_SERV,
  VEG_SERV,
  WGRAIN_SERV,
  LEGUMES_SERV,
  NUTS_SERV,
  FISH_SERV,
  REDPROC_MEAT_SERV,
  MONSATFAT_SERV,
  ALCOHOL_SERV
)
}
\arguments{
\item{SERV_DATA}{The raw data file that includes all the serving sizes of foods and nutrients}

\item{FRT_FRTJ_SERV}{The serving size of All fruits and 100\% fruit juices, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g); 1 cup fruit juice}

\item{VEG_SERV}{The serving size of All vegetables except potatoes and legumes, unit=0.5 c of vege; 1 cup of green leafy}

\item{WGRAIN_SERV}{The serving size of whole grains, including Whole-grain ready-to-eat cereals, cooked cereals, crackers, dark breads, brown rice, other grains, wheat germ, bran, popcorn, unit=1oz}

\item{LEGUMES_SERV}{The serving size of legumes, including Tofu, string beans, peas, beans, unit=1cup}

\item{NUTS_SERV}{The serving size of nuts, including Nuts, peanut butter, unit=1oz}

\item{FISH_SERV}{The serving size of all fish, including Fish and shrimp, breaded fish, unit=4oz}

\item{REDPROC_MEAT_SERV}{The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=serving (4 oz. unprocessed meat; 1.5 oz. processed meat)}

\item{MONSATFAT_SERV}{The serving size of the ratio of monounsaturated fat to saturated fat, unit=ratio}

\item{ALCOHOL_SERV}{The serving size of alcohol, unit=13g}
}
\value{
The MED index/score
}
\description{
Calculate the MED dietary index, Mediterranean, using given the serving sizes of foods and nutrients consumed per 1 day
}
\examples{
MED = function(SERV_DATA, FRT_FRTJ_SERV, VEG_SERV, WGRAIN_SERV, LEGUMES_SERV, NUTS_SERV,FISH_SERV, REDPROC_MEAT_SERV, MONSATFAT_SERV, ALCOHOL_SERV)
}
