% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\name{MED_NHANES_FPED}
\alias{MED_NHANES_FPED}
\title{MED_NHANES_FPED}
\usage{
MED_NHANES_FPED(FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
\arguments{
\item{FPED_PATH}{The file path for the FPED data. The file name should be like: fpre_dr1tot_1718.sas7bdat}

\item{NUTRIENT_PATH}{The file path for the NUTRIENT data. The file name should be like: DR1TOT_J.XPT}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT}
}
\value{
The MED and its component scores
}
\description{
Calculate the MED for the NHANES_FPED data (after 2005) within 1 step
}
\examples{
FPED_PATH = "/Users/james/Desktop/data/fpre_dr1tot_1718.sas7bdat"
NUTRIENT_PATH = "/Users/james/Desktop/data/DR1TOT_J.XPT"
DEMO_PATH = "/Users/james/Desktop/data/DEMO_J.XPT"
MED_NHANES_FPED (FPED_PATH, NUTRIENT_PATH, DEMO_PATH)
}
