% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietaryindex.R
\docType{data}
\name{STD_FOOD_FREQ}
\alias{STD_FOOD_FREQ}
\title{AHEI_BLOCK Calculation}
\format{
An object of class \code{numeric} of length 9.
}
\usage{
STD_FOOD_FREQ
}
\arguments{
\item{RAW_DATA}{The raw data file that includes results and raw responses of the dietary assessment}
}
\value{
The AHEI and its component scores
}
\description{
Calculate the AHEI dietary index for the Block FFQ (2013) per 1 day
}
\examples{
AHEI_BLOCK(RAW_DATA)
}
\keyword{datasets}
